Created basic leaderboard cell

+++ HackIllinois/ViewControllers/HIOnboardingViewController.swift
@@ -55,8 +55,8 @@ extension HIOnboardingViewController {
         setupUI()
         let viewAlphaValue = shouldDisplayAnimationOnNextAppearance ? 0.0 : 1.0
 
-        carouselView?.alpha = viewAlphaValue
-        getStartedButton.alpha = viewAlphaValue
+        carouselView?.alpha = CGFloat(viewAlphaValue)
+        getStartedButton.alpha = CGFloat(viewAlphaValue)
     }
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
